% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANScore.calculate.R
\name{PANScore.calculate}
\alias{PANScore.calculate}
\title{Calculate the PANoptosis related risk score}
\usage{
PANScore.calculate(exprset)
}
\arguments{
\item{expr}{log2 transformed TPM matrix derived from  normalized gene expression matrix from bulk samples; rows are genes, cols are samples.}
}
\value{
A dataframe containing two columns: continuous numerical variables with Riskscore, the ID of the samples.
}
\description{
#' This is some description of this function.
}
\examples{
\dontrun{
data("exprset.hovon")
RCDscore <- PANScore.calculate(exprset  = exprset.hovon)
}
}
\author{
Wei Zhang
}
